Avoid panic when u32 subtruction mades negative integer

+++ src/gbi/sound_processing.rs
@@ -10,5 +10,5 @@ pub fn pulse(phase: f64, duty: f64) -> i4 {
 }
 
 pub fn frequency_from_note_number(note_num: u16, note_num_440hz: u16) -> f64 {
-    440.0 * 2.0_f64.powf((note_num - note_num_440hz) as f64 / 12.0)
+    440.0 * 2.0_f64.powf((note_num as i32 - note_num_440hz as i32) as f64 / 12.0)
 }
